weather api complete

+++ lib/models/rest_models/AccuWeatherModels/geoposition/TimeZone.dart
@@ -1,7 +1,7 @@
 class TimeZone {
   String code;
   String name;
-  int gmtOffset;
+  double gmtOffset;
   bool isDaylightSaving;
   String nextOffsetChange;
